+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `set_power_amplification_ramp`, `set_pa_ramp` methods
 - `valid_rx_headers`, `valid_rx_packets` methods
 - `invert_iq` method and `InvertIQ` struct
+- `set_invert_iq` method
 
 ## [0.1.0] - 2026-03-07